+++ CHANGES.rst
@@ -48,6 +48,7 @@ v3.0.0 (xxxx-xx-xx)
 - Add ``to_number``.
 - Add ``underscore_case`` as alias of ``snake_case``.
 - Add ``unquote``.
+- Fix ``remove`` so that array is modified after callback iteration.
 - Fix ``trunc`` and ``truncate`` so that they handle texts shorter than the max string length correctly.
 - Make the following functions work with empty strings and ``None``: (**breaking change**)